add and use Platfrom
