+++ src/main/kotlin/com/anatawa12/mcLauncher/Profile.kt
@@ -4,6 +4,6 @@ data class Profile(
     val version: String,
     val gameDirPath: String,
     val jvmArguments: List<String> = listOf(),
-    val operatingSystem: OperatingSystem = OperatingSystem.current,
-    val appDataDirPath: String = operatingSystem.getAppDataDirPath()
+    val platform: Platform = Platform.current,
+    val appDataDirPath: String = platform.os.getAppDataDirPath()
 )
